Lift complexity of toggling ticker out of the preview component
