+++ public/src/components/channelManagement/bannerDesigns/StickyTopBar.tsx
@@ -1,22 +1,44 @@
-import React from 'react';
-import { Theme, Typography, makeStyles, Button } from '@material-ui/core';
+import React, { useState } from 'react';
+import {
+  Theme,
+  Typography,
+  makeStyles,
+  Button,
+  FormControlLabel,
+  Checkbox,
+} from '@material-ui/core';
 import EditIcon from '@material-ui/icons/Edit';
 import { LockDetails } from './LockDetails';
 import LockIcon from '@material-ui/icons/Lock';
 import CloseIcon from '@material-ui/icons/Close';
 import SaveIcon from '@material-ui/icons/Save';
 import { grey } from '@material-ui/core/colors';
-import { LockStatus } from '../helpers/shared';
+import { LockStatus, TickerCountType, TickerEndType, TickerName } from '../helpers/shared';
 import LiveSwitch from '../../shared/liveSwitch';
 import BannerVariantPreview from '../bannerTests/bannerVariantPreview';
 import { getDefaultVariant } from '../bannerTests/utils/defaults';
 import { BannerVariant } from '../../../models/banner';
 import { BannerDesign, Status } from '../../../models/bannerDesign';
 
-const buildVariantForPreview = (design: BannerDesign): BannerVariant => {
+const buildVariantForPreview = (design: BannerDesign, shouldShowTicker: boolean): BannerVariant => {
+  const tickerSettings = shouldShowTicker
+    ? {
+        countType: TickerCountType.money,
+        endType: TickerEndType.hardstop,
+        currencySymbol: '£',
+        copy: {
+          countLabel: 'contributions in May',
+          goalReachedPrimary: "We've met our goal - thank you!",
+          goalReachedSecondary: '',
+        },
+        name: TickerName.US,
+      }
+    : undefined;
+
   return {
     ...getDefaultVariant(),
     template: { designName: design.name },
+    tickerSettings,
   };
 };
 
@@ -83,6 +105,26 @@ interface Props {
   onStatusChange: (status: Status) => void;
 }
 
+interface TickerToggleProps {
+  shouldShowTicker: boolean;
+  setShouldShowTicker: (shouldShowTicker: boolean) => void;
+}
+
+const TickerToggle = ({ shouldShowTicker, setShouldShowTicker }: TickerToggleProps) => {
+  return (
+    <FormControlLabel
+      control={
+        <Checkbox
+          checked={shouldShowTicker}
+          onChange={() => setShouldShowTicker(!shouldShowTicker)}
+          color="primary"
+        />
+      }
+      label={'Show ticker?'}
+    />
+  );
+};
+
 const StickyTopBar: React.FC<Props> = ({
   name,
   design,
@@ -94,6 +136,8 @@ const StickyTopBar: React.FC<Props> = ({
   onStatusChange,
 }: Props) => {
   const classes = useStyles();
+
+  const [shouldShowTicker, setShouldShowTicker] = useState<boolean>(false);
 
   return (
     <header className={classes.container}>
@@ -158,9 +202,14 @@ const StickyTopBar: React.FC<Props> = ({
             </>
           )}
           <BannerVariantPreview
-            variant={buildVariantForPreview(design)}
+            variant={buildVariantForPreview(design, shouldShowTicker)}
             design={design}
-            shouldShowTickerToggle={true}
+            controls={
+              <TickerToggle
+                shouldShowTicker={shouldShowTicker}
+                setShouldShowTicker={setShouldShowTicker}
+              />
+            }
           />
         </div>
       </div>
